HW6 Added FirebaseManager
